pyproject.toml
[tool.poetry]
name = "modelib"
version = "0.1.0"
description = ""
authors = ["Gabriel Guarisa <gabrielguarisa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
joblib = "^1.3.2"
uvicorn = {extras = ["standard"], version = "^0.26.0"}
fastapi = "^0.109.0"
prometheus-fastapi-instrumentator = "^6.1.0"
python-json-logger = "^2.0.7"
pandas = "*"
scikit-learn = "*"
python-slugify = "^8.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.13"
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
